+++ src/main/java/com/ralf/NewDiverStore/producer/controller/ProducerAdminViewController.java
@@ -1,12 +1,18 @@
 package com.ralf.NewDiverStore.producer.controller;
 
+import com.ralf.NewDiverStore.common.dto.Message;
 import com.ralf.NewDiverStore.producer.domain.model.Producer;
 import com.ralf.NewDiverStore.producer.service.ProducerService;
+import com.ralf.NewDiverStore.product.domain.model.Product;
 import com.ralf.NewDiverStore.product.service.ProductService;
+import jakarta.validation.Valid;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import java.util.List;
 import java.util.UUID;
 
 
@@ -31,8 +37,14 @@ public class ProducerAdminViewController {
     }
 
     @GetMapping("{id}/delete")
-    public String deleteView(@PathVariable UUID id) {
-        producerService.deleteProducer(id);
+    public String deleteView(@PathVariable UUID id, RedirectAttributes ra) {
+        List<Product> productsList = productService.getProductsByProducerId(id);
+        if (productsList.isEmpty()) {
+            producerService.deleteProducer(id);
+            ra.addFlashAttribute("message", Message.info("Producer removed"));
+        } else {
+            ra.addFlashAttribute("message", Message.error("Producer cannot be removed. Please delete related products first"));
+        }
         return "redirect:/admin/producers";
     }
 
@@ -43,8 +55,26 @@ public class ProducerAdminViewController {
     }
 
     @PostMapping("{id}/edit")
-    public String edit(@ModelAttribute("producer") Producer producer, @PathVariable UUID id) {
-        producerService.updateProducer(id, producer);
+    public String edit(@PathVariable UUID id,
+                       @Valid @ModelAttribute("producer") Producer producer,
+                       BindingResult bindingResult,
+                       RedirectAttributes ra,
+                       Model model
+    ) {
+        if (bindingResult.hasErrors()) {
+            model.addAttribute("producer", producer);
+            model.addAttribute("message", Message.error("Data writing error"));
+            return "admin/producer/edit";
+        }
+
+        try {
+            producerService.updateProducer(id, producer);
+            ra.addFlashAttribute("message", Message.info("Producer updated"));
+        } catch (Exception e) {
+            model.addAttribute("producer", producer);
+            model.addAttribute("message", Message.error("Unknown data writing error"));
+            return "admin/producer/edit";
+        }
         return "redirect:/admin/producers";
     }
 
@@ -62,10 +92,30 @@ public class ProducerAdminViewController {
     }
 
     @PostMapping("add")
-    public String add(@ModelAttribute("producer")Producer producer){
+    public String add(@ModelAttribute("producer") Producer producer) {
         producerService.createProducer(producer);
         return "redirect:/admin/producers";
     }
 
 
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
